Revised waterlog
